refactor: move types to each directory
