+++ src/api/services/github/types.ts
@@ -1,0 +1,52 @@
+export type GithubTeam = {
+  name: string;
+  id: number;
+  node_id: string;
+  slug: string;
+  description: string;
+  privacy: string;
+  notification_setting: string;
+  url: string;
+  html_url: string;
+  members_url: string;
+  repositories_url: string;
+  permission: string;
+  parent: null;
+};
+
+export type GithubMember = {
+  login: string;
+  id: number;
+  node_id: string;
+  avatar_url: string;
+  gravatar_id: string;
+  url: string;
+  html_url: string;
+  followers_url: string;
+  following_url: string;
+  gists_url: string;
+  starred_url: string;
+  subscriptions_url: string;
+  organizations_url: string;
+  repos_url: string;
+  events_url: string;
+  received_events_url: string;
+  type: string;
+  user_view_type: string;
+  site_admin: boolean;
+};
+
+export type GithubTreeResponse = {
+  sha: string;
+  url: string;
+  tree: GithubTree[];
+};
+
+export type GithubTree = {
+  path: string;
+  mode: string;
+  type: string;
+  size: number;
+  sha: string;
+  url: string;
+};
