+++ src/api/utils/types.ts
@@ -1,0 +1,1 @@
+export type HeaderTuple = [string, string];
